requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
requests==2.32.3
python-dotenv==1.0.1

# ── Web & Dashboard ───────────────────────────────────────────────────────────
streamlit==1.36.0
plotly==5.22.0

# ── AI Client ────────────────────────────────────────────────────────────────
openai==1.35.3

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.2.2
openpyxl==3.1.4

# ── Utilities ─────────────────────────────────────────────────────────────────
urllib3==2.2.1
